+++ c2pa/core.py
@@ -16,33 +16,37 @@
 # along with pyc2pa.  If not, see <http://www.gnu.org/licenses/>.
 
 import hashlib
+import uuid
+import datetime
 
 import multibase
 import multihash
 import pyexiv2
+import cbor
+import pytz
 
-from Crypto.Hash import SHA256
-from Crypto.PublicKey import RSA
-from Crypto.Signature import PKCS1_v1_5
+from cryptography.hazmat.backends import default_backend
+from cryptography import x509
+from cryptography.hazmat.primitives.serialization import Encoding
+from cryptography.hazmat.primitives import hashes, serialization
+from cryptography.hazmat.primitives.asymmetric import padding
 
-from cai.jumbf import ContentBox
-from cai.jumbf import DescriptionBox
-from cai.jumbf import SuperBox
+from c2pa.jumbf import ContentBox
+from c2pa.jumbf import DescriptionBox
+from c2pa.jumbf import SuperBox
 
-from cai.jumbf import json_to_bytes
+from c2pa.jumbf import json_to_cbor_bytes
 
-from endesive import plain
-
-'''Implementation of CAI Whitepaper
+'''Implementation of C2PA Whitepaper
 Content Authenticity Initiative
 '''
 
 Cai_content_types = {
-    'claim_block'    : '6361636200110010800000aa00389b71', # noqa E203
-    'store'          : '6361737400110010800000aa00389b71', # noqa E203
-    'assertion_store': '6361617300110010800000aa00389b71',
-    'claim'          : '6361636c00110010800000aa00389b71', # noqa E203
-    'claim_signature': '6361736700110010800000aa00389b71',
+    'manifest_block'    : '6332706100110010800000aa00389b71',
+    'manifest'          : '63326D6100110010800000aa00389b71',
+    'assertion_store'   : '6332617300110010800000aa00389b71',
+    'claim'             : '6332636C00110010800000aa00389b71',
+    'claim_signature'   : '6332637300110010800000aa00389b71',
 }
 
 
@@ -83,11 +87,16 @@ def encode_hashlink(binary_content, codec='base64', to_hexstr=False):
         return mb
 
 
-def insert_xmp_key(data_bytes, store_label='cai/cb.starling_1'):
+def compute_hash(binary_content):
+    m = hashlib.sha256(binary_content)
+    return m.digest()
+
+
+def insert_xmp_key(data_bytes, manifest_label):
     metadata = pyexiv2.ImageMetadata.from_buffer(data_bytes)
     metadata.read()
-    metadata['Xmp.dcterms.provenance'] = pyexiv2.XmpTag('Xmp.dcterms.provenance',
-                                                        'self#jumbf=cai/{}/cai.claim'.format(store_label))
+    metadata['Xmp.dcterms.provenance'] = pyexiv2.XmpTag(
+        'Xmp.dcterms.provenance', 'self#jumbf=c2pa/{}/c2pa.claim'.format(manifest_label))
     metadata.write()
     return metadata.buffer
 
@@ -101,159 +110,132 @@ def get_xmp_tag(data_bytes, tag='Xmp.dcterms.provenance'):
         return ''
 
 
-class CaiAssertionStore(SuperBox):
+class C2paAssertionStore(SuperBox):
     def __init__(self, assertions):
-        super(CaiAssertionStore, self).__init__()
+        super(C2paAssertionStore, self).__init__()
         self.description_box = DescriptionBox(
                                    content_type=Cai_content_types['assertion_store'],
-                                   label='cai.assertions')
+                                   label='c2pa.assertions')
         self.content_boxes = assertions
 
 
-class CaiClaim(SuperBox):
-    def __init__(self, assertion_store,
-                 store_label='cb.starling_1',
-                 recorder='Starling Capture using Numbers Protocol',
-                 parent_claim=''):
-        super(CaiClaim, self).__init__()
+class C2paClaim(SuperBox):
+    def __init__(self,
+                 assertion_store,
+                 manifest_label,
+                 media_name,
+                 recorder='Starling Capture using Numbers Protocol'):
+        super(C2paClaim, self).__init__()
         self.description_box = DescriptionBox(
-                                   content_type=Cai_content_types['claim'],
-                                   label='cai.claim')
-        content_box = ContentBox()
-        content_box.payload = json_to_bytes(
+                                    content_type=Cai_content_types['claim'],
+                                    label='c2pa.claim')
+        content_box = ContentBox(t_box_type='cbor')
+        content_box.payload = json_to_cbor_bytes(
             self.create_claim(assertion_store,
-                              store_label=store_label,
+                              manifest_label,
                               recorder=recorder,
-                              parent_claim=parent_claim)
+                              media_name=media_name)
         )
         self.content_boxes.append(content_box)
 
-    def create_claim(self, assertion_store,
-                     store_label='cb.starling_1',
-                     recorder='Starling Capture',
-                     parent_claim=''):
+    def create_claim(self,
+                     assertion_store,
+                     manifest_label,
+                     media_name,
+                     recorder='Starling Capture'):
         '''Create a Claim JSON object
         '''
         claim = {}
-        claim['assertions'] = [
-            'self#jumbf=cai/{store_label}/cai.assertions/{assertion_label}?hl={hashlink}'.format(
-                store_label=store_label,
+        claim['dc:title'] = media_name
+        claim['dc:format'] = 'image/jpeg'
+        claim['instanceID'] = 'xmp:fakeid:4124fae1-1da7-4a3f-95c8-d8ae071bd048'
+        claim['claim_generator'] = recorder
+        claim['signature'] = 'self#jumbf=c2pa/{}/c2pa.signature'.format(
+            manifest_label)
+        claim['assertions'] = [{
+            'url': 'self#jumbf=c2pa/{manifest_label}/c2pa.assertions/{assertion_label}'
+            .format(
+                manifest_label=manifest_label,
                 assertion_label=assertion.description_box.db_label,
-                hashlink=encode_hashlink(assertion.content_boxes[0].convert_bytes()[8:], to_hexstr=True)
-            )
-            for assertion in assertion_store.content_boxes
-        ]
-        claim['asset_hashes'] = Claim_asset_hashes_mockup
-        claim['recorder'] = recorder
-        claim['signature'] = 'self#jumbf=cai/{}/cai.signature'.format(store_label)
-        if parent_claim != '':
-            claim['parent_claim'] = parent_claim
+            ),
+            'alg': 'sha256',
+            'hash': compute_hash(assertion.content_boxes[0].convert_bytes()[8:]),
+        } for assertion in assertion_store.content_boxes]
+        claim['alg'] = 'sha256'
         return claim
 
 
-class CaiClaimEndesiveSignature(SuperBox):
-    def __init__(self, claim, key):
-        super(CaiClaimEndesiveSignature, self).__init__()
+class C2paClaimSignature(SuperBox):
+    def __init__(self, claim, key, certificate):
+        super(C2paClaimSignature, self).__init__()
         self.description_box = DescriptionBox(
-                                    content_type=Cai_content_types['claim_signature'],
-                                    label='cai.signature')
-        content_box = ContentBox(t_box_type='uuid')
-        content_box.payload = self.create_endesive_signature(claim, key)
+            content_type=Cai_content_types['claim_signature'],
+            label='c2pa.signature')
+        content_box = ContentBox(t_box_type='cbor')
+        content_box.payload = self.create_signature(claim, key, certificate)
         self.content_boxes.append(content_box)
 
-    def create_endesive_signature(self, claim, key):
-        uuid = Cai_content_types['claim_signature']
-        data = json_to_bytes(claim)
-        signature = plain.sign(data, key[0], key[1], key[2], 'sha256', attrs=True)
-        payload = bytes.fromhex(uuid) + signature
-        return payload
-
-
-class CaiClaimCMSSignature(SuperBox):
-    def __init__(self, claim, key):
-        super(CaiClaimCMSSignature, self).__init__()
-        self.description_box = DescriptionBox(
-                                    content_type=Cai_content_types['claim_signature'],
-                                    label='cai.signature')
-        content_box = ContentBox(t_box_type='uuid')
-        content_box.payload = self.create_cms_signature(claim, key)
-        self.content_boxes.append(content_box)
-
-    def generate_signature(self, data, key):
-        h = SHA256.new(data)
-        rsa = RSA.importKey(key)
-        signer = PKCS1_v1_5.new(rsa)
-        signature = signer.sign(h)
-        return signature
-
-    def create_cms_signature(self, claim, key):
-        uuid = Cai_content_types['claim_signature']
-        data = json_to_bytes(claim)
-        signature = self.generate_signature(data, key)
-        payload = bytes.fromhex(uuid) + signature
-        return payload
-
-
-class CaiClaimSignature(SuperBox):
-    def __init__(self):
-        super(CaiClaimSignature, self).__init__()
-        self.description_box = DescriptionBox(
-                                   content_type=Cai_content_types['claim_signature'],
-                                   label='cai.signature')
-        content_box = ContentBox(t_box_type='uuid')
-        content_box.payload = self.create_signature()
-        self.content_boxes.append(content_box)
-
-    def create_signature(self):
+    def create_signature(self, claim, key, certificate):
         '''Create a Claim Signature payload in bytes.
         '''
-        uuid = Cai_content_types['claim_signature']
-        signature = 'signature placeholder:cb.starling_1'
-        padding = b'\x20' * (100 - len(signature))
-        payload = bytes.fromhex(uuid) + signature.encode('utf-8') + padding
+        # -37 stands for PS256 (RSASSA-PSS using SHA-256 and MGF1 with SHA-256)
+        phdr = cbor.dumps({1: -37})
+        cert = x509.load_pem_x509_certificate(certificate, default_backend())
+        uhdr = {
+            'x5chain': cert.public_bytes(Encoding.DER),
+            'temp_signing_time': str(datetime.datetime.now(pytz.utc)),
+        }
+
+        private_key = serialization.load_pem_private_key(key, password=None)
+        sig_structure_data = cbor.dumps(['Signature1', phdr, b'' b'', claim])
+        sig_structure_data = b'\x84' + sig_structure_data[1:]
+        signature = private_key.sign(
+            sig_structure_data,
+            padding.PSS(
+                mgf=padding.MGF1(hashes.SHA256()),
+                salt_length=32
+            ),
+            hashes.SHA256()
+        )
+
+        payload = None
+        message = [phdr, uhdr, payload, signature]
+        tag = cbor.Tag(18, message)
+        cose_tag = cbor.dumps(tag)
+
+        pad = b'\x00' * (4096 - len(cose_tag))
+        payload = cose_tag + pad
         return payload
 
 
-class CaiStore(SuperBox):
-    def __init__(self, label='cb.starling_1',
+class C2paManifest(SuperBox):
+    def __init__(self,
+                 media_name,
+                 provider='numbersprotocol',
                  assertions=[],
                  recorder='Starling Capture',
-                 parent_claim='',
-                 key=[],
-                 sig='cms'):
-        super(CaiStore, self).__init__()
+                 key='',
+                 certificate=''):
+        super(C2paManifest, self).__init__()
+        self.manifest_label = '{}:urn:uuid:{}'.format(provider, uuid.uuid4())
         self.description_box = DescriptionBox(
-                                   content_type=Cai_content_types['store'],
-                                   label=label)
-        self.assertion_store = CaiAssertionStore(assertions)
-        self.claim = CaiClaim(self.assertion_store, recorder=recorder, parent_claim=parent_claim)
-        if len(key) == 0:
-            self.signature = CaiClaimSignature()
-        else:
-            if sig == 'cms':
-                self.signature = CaiClaimCMSSignature(
-                    self.claim.create_claim(
-                        self.assertion_store,
-                        recorder=recorder,
-                        parent_claim=parent_claim),
-                    key)
-            elif sig == 'endesive':
-                self.signature = CaiClaimEndesiveSignature(
-                    self.claim.create_claim(
-                        self.assertion_store,
-                        recorder=recorder,
-                        parent_claim=parent_claim),
-                    key)
-            else:
-                self.signature = CaiClaimSignature()
+            content_type=Cai_content_types['manifest'],
+            label=self.manifest_label)
+        self.assertion_store = C2paAssertionStore(assertions)
+        self.claim = C2paClaim(self.assertion_store,
+                               self.manifest_label,
+                               media_name,
+                               recorder=recorder)
+        self.signature = C2paClaimSignature(self.claim.content_boxes[0].payload, key, certificate)
+
         self.content_boxes.append(self.assertion_store)
         self.content_boxes.append(self.claim)
         self.content_boxes.append(self.signature)
 
 
-class CaiClaimBlock(SuperBox):
+class C2paManifestBlock(SuperBox):
     def __init__(self):
-        super(CaiClaimBlock, self).__init__()
+        super(C2paManifestBlock, self).__init__()
         self.description_box = DescriptionBox(
-                                   content_type=Cai_content_types['claim_block'],
-                                   label='cai')
+                                   content_type=Cai_content_types['manifest_block'],
+                                   label='c2pa')